Fix tests that previously assumed float32

+++ tensorly/tenalg/tests/test_proximal.py
@@ -22,6 +22,7 @@ from ..proximal import (
 from ...testing import assert_, assert_array_equal, assert_array_almost_equal
 from tensorly import tensor_to_vec, truncated_svd
 import pytest
+import tensorly as tl
 
 # Author: Jean Kossaifi
 skip_tensorflow = pytest.mark.skipif(
@@ -82,9 +83,9 @@ def test_simplex():
 
 def test_normalized_sparsity():
     """Test for normalized_sparsity operator"""
-    tensor = T.tensor([2, 3, 4])
+    tensor = T.tensor([2, 3, 4], dtype=tl.float32)
     res = normalized_sparsity_prox(tensor, 2)
-    true_res = T.tensor([0, 0.6, 0.8])
+    true_res = T.tensor([0, 0.6, 0.8], dtype=tl.float32)
     assert_array_almost_equal(true_res, res)
 
 
@@ -120,15 +121,15 @@ def test_unimodality():
 
 def test_l2_prox():
     """Test for l2 prox operator"""
-    tensor = T.tensor([2, 4, 4])
+    tensor = T.tensor([2, 4, 4], dtype=tl.float32)
     res = l2_prox(tensor, 3)
-    true_res = T.tensor([1, 2, 2])
+    true_res = T.tensor([1, 2, 2], dtype=tl.float32)
     assert_array_almost_equal(true_res, res)
 
 
 def test_squared_l2_prox():
     """Test for squared l2 prox operator"""
-    tensor = T.tensor([3, 6, 9])
+    tensor = T.tensor([3, 6, 9], dtype=tl.float32)
     res = l2_square_prox(tensor, 0.5)
     true_res = T.tensor([1.5, 3, 4.5])
     assert_array_almost_equal(true_res, res)
@@ -182,10 +183,10 @@ def test_soft_thresholding():
 
 def test_svd_thresholding():
     """Test for singular_value_thresholding operator"""
-    U = T.tensor([[1, 0, 0], [0, 1, 0], [0, 0, 1]])
+    U = T.tensor([[1, 0, 0], [0, 1, 0], [0, 0, 1]], dtype=tl.float32)
     singular_values = T.tensor([0.4, 2.1, -2])
     tensor = T.dot(U, T.reshape(singular_values, (-1, 1)) * T.transpose(U))
-    shrinked_singular_values = T.tensor([0, 1.6, -1.5])
+    shrinked_singular_values = T.tensor([0, 1.6, -1.5], dtype=tl.float32)
     true_res = T.dot(U, T.reshape(shrinked_singular_values, (-1, 1)) * T.transpose(U))
     res = svd_thresholding(tensor, 0.5)
     assert_array_almost_equal(true_res, res)
